move auto complete to it's own module
-fix search overwriting dot
-preparing to make vim it's own module
